Refactored utils to separate modules
