+++ src/utils/index.js
@@ -1,0 +1,9 @@
+export {getRandomInteger} from "./random-integer";
+export {getRandomArrayItem} from "./random-array-item";
+export {getRandomDate} from "./random-date";
+export {getRandomBoolean} from "./random-boolean";
+export {formatDate} from "./format-date";
+export {formatNumber} from "./format-number";
+export {createElement} from "./create-element";
+export {render, RenderPosition} from "./render";
+
