+++ src/utils/random-boolean.js
@@ -1,0 +1,7 @@
+/**
+ * Returns random boolean value (true or false)
+ * @return {Boolean} - random true or false
+ */
+const getRandomBoolean = () => Math.random() > 0.5;
+
+export {getRandomBoolean};
